feat(story): split up stories folder
