+++ .vscode/settings.json
@@ -1,6 +1,7 @@
 {
     "editor.fontFamily": "",
     "cSpell.words": [
-        "cormorantinfant"
+        "cormorantinfant",
+        "Tablist"
     ]
 }
