+++ app/components/Manager/Manager.tsx
