+++ app/components/TabList/index.ts
@@ -1,0 +1,1 @@
+export * from './Tablist'
